+++ jsconfig.json
@@ -7,6 +7,7 @@
       "@/pages/*": ["components/pages/*"],
       "@/utils/*": ["utils/*"],
       "@/constants/*": ["constants/*"],
+      "@/components/*": ["components/*"],
       "@/test/*": ["test/*"],
       "@/i18n/*": ["i18n/*"],
       "@/layouts/*": ["components/layouts/*"],
