--- src/components/pages/ContextProvider.js
@@ -1,22 +1,0 @@
-import PropTypes from 'prop-types';
-
-const ContextProvider = ({ providers, children }) => {
-  return providers.reverse().reduce((AccumulatedProviders, current) => {
-    const [CurrentProvider, currentProps] = Array.isArray(current)
-      ? current
-      : [current, {}];
-    // eslint-disable-next-line react/prop-types
-    return (
-      <CurrentProvider {...currentProps}>
-        {AccumulatedProviders}
-      </CurrentProvider>
-    );
-  }, children);
-};
-
-ContextProvider.propTypes = {
-  providers: PropTypes.array,
-  children: PropTypes.node,
-};
-
-export { ContextProvider };
