+++ src/components/pages/_app.js
@@ -14,7 +14,25 @@ import '@fortawesome/fontawesome-svg-core/styles.css';
 import { GlobalStyle } from '@/styles/GlobalStyle';
 
 import Layout from '@/layouts/index';
-import { ContextProvider } from '@/pages/ContextProvider';
+
+const ContextProvider = ({ providers, children }) => {
+  return providers.reverse().reduce((AccumulatedProviders, current) => {
+    const [CurrentProvider, currentProps] = Array.isArray(current)
+      ? current
+      : [current, {}];
+    // eslint-disable-next-line react/prop-types
+    return (
+      <CurrentProvider {...currentProps}>
+        {AccumulatedProviders}
+      </CurrentProvider>
+    );
+  }, children);
+};
+
+ContextProvider.propTypes = {
+  providers: PropTypes.array,
+  children: PropTypes.node,
+};
 
 config.autoAddCss = false;
